done part 4

+++ part4/controllers/blogs.js
@@ -1,6 +1,7 @@
 const blogRouter = require('express').Router();
 const Blog = require('../models/blog');
 const User = require('../models/user');
+const jwt = require('jsonwebtoken');
 
 // get all blogs
 blogRouter.get('/', async (request, response) => {
@@ -23,7 +24,14 @@ blogRouter.get('/:id', async (request, response) => {
 //post a blog
 blogRouter.post('/', async (request, response) => {
   const body = request.body;
-  const user = await User.findById(body.userId);
+  const token = request.token;
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' });
+  }
+  const user = request.user;
+
   const blog = new Blog({
     title: body.title,
     author: body.author,
@@ -41,8 +49,19 @@ blogRouter.post('/', async (request, response) => {
 
 // delete a blog
 blogRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  const token = request.token;
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' });
+  }
+  const user = request.user;
+  const blog = await Blog.findById(request.params.id);
+  if (blog.user._id.toString() === user._id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  } else {
+    return response.status(401).json({ error: 'wrong user' });
+  }
 });
 
 //update by id
